fix: request_parser 함수로 분리
